requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
docopt==0.6.2
idna==3.4
mysqlclient==2.2.0
oauthlib==3.2.2
pipreqs==0.4.13
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
wheel==0.41.3
yarg==0.1.9
